Added another check to the music player for proper memory allocation.  If there is insufficient memory to allocate a track buffer, the music player won't run.
Added an explicit call to System.gc() when the music player loads a new song, as garbage collection per the default Android heuristics was sometimes inadequate to meet the dynamic memory demands of the game, especially when completing many levels in very rapid succession.

Improved the efficiency of the network protocol significantly for WiFi multiplayer with improved handling of the timing of player status messages.